feat: improve structure of convex functions
